% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_doi_list_altmetrics}
\alias{annotate_doi_list_altmetrics}
\title{Annotate a list of DOIs using Altmetric}
\usage{
annotate_doi_list_altmetrics(doi_list)
}
\arguments{
\item{doi_list}{A \emph{list} of DOI character strings (as expected by
\code{purrr::pmap_df}).}
}
\value{
A data frame of Altmetric scores and attention metrics.
}
\description{
Retrieves Altmetric attention scores and social-media metrics for the
supplied DOIs.
}
\examples{
\dontrun{
art_test <- get_article_most_recent_table("Zeitgeber")
dois <- unique(unlist(stringr::str_match_all(
  art_test$`*`, "10\\\\.\\\\d{4,9}/[-._;()/:a-z0-9A-Z]+"
)))
annotate_doi_list_altmetrics(list(dois[1:3]))
}
}
