% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_doi_to_bibtex_cross_ref}
\alias{annotate_doi_to_bibtex_cross_ref}
\title{Retrieve BibTeX entries for a list of DOIs via CrossRef}
\usage{
annotate_doi_to_bibtex_cross_ref(doi_list)
}
\arguments{
\item{doi_list}{Character vector of DOIs.}
}
\value{
A list of BibTeX character strings, one per DOI.
}
\description{
Retrieve BibTeX entries for a list of DOIs via CrossRef
}
\examples{
\donttest{
annotate_doi_to_bibtex_cross_ref("10.1038/nature12373")
}
}
