% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_isbn_list_altmetrics}
\alias{annotate_isbn_list_altmetrics}
\title{Annotate a list of ISBNs using Altmetric}
\usage{
annotate_isbn_list_altmetrics(isbn_list)
}
\arguments{
\item{isbn_list}{A \emph{list} of ISBN character strings.}
}
\value{
A data frame of Altmetric scores for the supplied ISBNs.
}
\description{
Annotate a list of ISBNs using Altmetric
}
\examples{
\dontrun{
# Requires the optional 'rAltmetric' package (not on CRAN):
#   remotes::install_github('ropensci/rAltmetric')
annotate_isbn_list_altmetrics(list(c("9780156031356")))
}
}
