% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{extract_citations}
\alias{extract_citations}
\title{Extract all Citation Style 1 templates from wikitext}
\usage{
extract_citations(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
Character vector of matched CS1 templates.
}
\description{
Extract all Citation Style 1 templates from wikitext
}
\examples{
\donttest{
art <- get_article_most_recent_table("Zeitgeber")
extract_citations(art$`*`)
}
}
