% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{extract_wikihypelinks}
\alias{extract_wikihypelinks}
\title{Extract Wikipedia hyperlinks from wikitext}
\usage{
extract_wikihypelinks(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
Character vector of matched \code{[[...]]}-style links.
}
\description{
Extract Wikipedia hyperlinks from wikitext
}
\examples{
extract_wikihypelinks("See [[Zeitgeber|the article]] and [[Sleep]].")
}
