% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{get_anyCount}
\alias{get_anyCount}
\title{Count matches of an arbitrary regular expression in wikitext}
\usage{
get_anyCount(art_text, regexp)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}

\item{regexp}{A regular expression string.}
}
\value{
Integer count of matches.
}
\description{
Count matches of an arbitrary regular expression in wikitext
}
\examples{
get_anyCount("foo bar baz foo", "foo")
}
