% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_viz.R
\name{get_closest_date}
\alias{get_closest_date}
\title{Find the closest date in a vector to a reference date}
\usage{
get_closest_date(date_in, date_vect)
}
\arguments{
\item{date_in}{A \code{Date} or date-coercible value.}

\item{date_vect}{A vector of \code{Date} values.}
}
\value{
The element of \code{date_vect} closest in time to \code{date_in}.
}
\description{
Utility used for aligning paper publication dates to Wikipedia revision
snapshots.
}
\examples{
dates <- as.Date(c("2020-01-01", "2020-06-01", "2021-01-01"))
get_closest_date(as.Date("2020-04-15"), dates)
}
