% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{get_doi_count}
\alias{get_doi_count}
\title{Count DOIs in wikitext}
\usage{
get_doi_count(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
Integer count of matched DOIs.
}
\description{
Count DOIs in wikitext
}
\examples{
get_doi_count("See 10.1038/nature12373 and 10.1016/j.cell.2020.01.001.")
}
