% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{get_hyperlinkCount}
\alias{get_hyperlinkCount}
\title{Count Wikipedia hyperlinks in wikitext}
\usage{
get_hyperlinkCount(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
Integer count of \code{[[...]]}-style links.
}
\description{
Count Wikipedia hyperlinks in wikitext
}
\examples{
get_hyperlinkCount("[[Article one]] and [[Article two|display]]")
}
