% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{get_ISBN_count}
\alias{get_ISBN_count}
\title{Count ISBNs in wikitext}
\usage{
get_ISBN_count(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
Integer count of matched ISBNs.
}
\description{
Count ISBNs in wikitext
}
\examples{
get_ISBN_count("Book isbn=978-0-06-112008-4")
}
