% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{get_refCount}
\alias{get_refCount}
\title{Count \code{<ref>} tags in wikitext}
\usage{
get_refCount(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
Integer count of reference tags.
}
\description{
Count \code{<ref>} tags in wikitext
}
\examples{
get_refCount("<ref>one</ref> text <ref name='r2'>two</ref>")
}
