% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_parsing.R
\name{get_source_type_counts}
\alias{get_source_type_counts}
\title{Count citations by CS1 source type}
\usage{
get_source_type_counts(art_text)
}
\arguments{
\item{art_text}{Character string of raw wikitext.}
}
\value{
A data frame with columns \code{cite_type} and \code{Freq}, or
  \code{NA} if no citations are found.
}
\description{
Count citations by CS1 source type
}
\examples{
\dontrun{
art <- get_article_most_recent_table("Zeitgeber")
get_source_type_counts(art$`*`)
}
}
