% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_history.R
\name{get_subcat_multiple}
\alias{get_subcat_multiple}
\title{Retrieve subcategories for multiple Wikipedia categories}
\usage{
get_subcat_multiple(catlist, replecement = "_")
}
\arguments{
\item{catlist}{Character vector of category names.}

\item{replecement}{Character used to replace spaces (default: \code{"_"}).}
}
\value{
A combined data frame of subcategory metadata.
}
\description{
Retrieve subcategories for multiple Wikipedia categories
}
\examples{
\dontrun{
get_subcat_multiple(c("Category:Biology", "Category:Medicine"))
}
}
