% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wiki_cache_dir}
\alias{wiki_cache_dir}
\title{Return (and create if needed) the wikilite user cache directory}
\usage{
wiki_cache_dir()
}
\value{
Character string path to the cache directory.
}
\description{
Return (and create if needed) the wikilite user cache directory
}
